Create blumenstock.md
The possible data science developments that Joshua Blumenstock discusses in his article are utopian and their and appear to be a positive attempt and furthering international development. There is definitely promise in the ways to target those of need in order to get the help they need. Proper utilization promises getting resources to those in need in a timely cost-effective manner. Collaboration among systems appears to promote a more flawless tactic for pinpointing areas that seek development. As good as everything sounds, there are some pitfalls. Blumenstock elaborates on at least four including effects, lack of validity, bias, and lack of regulation. It's possible that efforts remain out of reach for the most vulnerable and impact those who do not necessarily need the aid. There is always room for error when it comes to digital data collection. Algorithms can often be biased and disregard certain outliers and marginalized groups. Many groups are unwilling to completely open themselves up through technology as a need for privacy and protection is at stake. 
The statements made by students are all very inquisitive and have meanings to consider. Anna’s statement has a lot of truth to it. Not just in this scenario but with many, sometimes plans with good intentions tend to miss the mark and do more harm than good. This relates to the unanticipated effects that Blumenstock mentioned. Nira’s message about transparency is true but not necessarily a make or break point in both data based and human based issues. I do think it is important for those creating, collecting, and distributing data in order to further international development should remain transparent about their efforts. If researchers, analysts, etc. were more open, individuals may be more willing to open and problems such as the lack of regulation that Blumenstock mentioned could be solved. Kayla’s quote is a lot to unpack. There is a lot of promise in what is being proposed but it comes with possible consequences such as improper representation and marginalization. It will be a challenge to create a proper system that will account for those not engaging in certain aspects like social media leaving them out of some of the proposed solutions. It can be hard to relate a topic with no face, and numbers and facts like data science and something like human development that introduces morals and ethics. All the statements made by the students are great points and leave a lot up for discussion.